requirements.txt
aiogram
google-generativeai
python-telegram-bot